requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
sqlalchemy
groq
pydantic
